+++ app/main/views/dashboard.py
@@ -30,12 +30,11 @@ from app.utils import (
     DELIVERED_STATUSES,
     FAILURE_STATUSES,
     REQUESTED_STATUSES,
-    generate_next_dict,
-    generate_previous_dict,
     get_current_financial_year,
     service_has_permission,
 )
 from app.utils.csv import Spreadsheet
+from app.utils.pagination import generate_next_dict, generate_previous_dict
 from app.utils.user import user_has_permissions
