+++ app/utils/__init__.py
@@ -55,32 +55,6 @@ def redirect_to_sign_in(f):
     return wrapped
 
 
-def get_page_from_request():
-    if 'page' in request.args:
-        try:
-            return int(request.args['page'])
-        except ValueError:
-            return None
-    else:
-        return 1
-
-
-def generate_previous_dict(view, service_id, page, url_args=None):
-    return generate_previous_next_dict(view, service_id, page - 1, 'Previous page', url_args or {})
-
-
-def generate_next_dict(view, service_id, page, url_args=None):
-    return generate_previous_next_dict(view, service_id, page + 1, 'Next page', url_args or {})
-
-
-def generate_previous_next_dict(view, service_id, page, title, url_args):
-    return {
-        'url': url_for(view, service_id=service_id, page=page, **url_args),
-        'title': title,
-        'label': 'page {}'.format(page)
-    }
-
-
 def get_help_argument():
     return request.args.get('help') if request.args.get('help') in ('1', '2', '3') else None
